+++ emacs/merlin.el
@@ -209,6 +209,12 @@ kill ring and C-d destructures the expression."
 `merlin-allow-sit-for' is `t'."
   :group 'merlin :type 'boolean)
 
+(defcustom merlin-construct-with-local-values nil
+  "If non-nil, `merlin-construct' includes values in the local environment.
+
+Otherwise, `merlin-construct' only includes constructors."
+  :group 'merlin :type 'boolean)
+
 (defalias 'merlin-find-file 'find-file-other-window
   "The function called when merlin try to open a file (doesn't apply to
 merlin-locate, see `merlin-locate-in-new-window').")
@@ -251,8 +257,6 @@ The association list can contain the following optional keys:
 ;;;;;;;;;;;;;;;;;;;;;;;;
 ;; Internal variables ;;
 ;;;;;;;;;;;;;;;;;;;;;;;;
-
-(defvar merlin-opam-bin-path nil)
 
 ;; If user did not specify its merlin-favourite-caml-mode, try to guess it from
 ;; the buffer being edited
@@ -506,7 +510,13 @@ argument (lookup appropriate binary, setup logging, pass global settings)"
   ;; Really start process
   (let ((binary      (merlin-command))
         ;; (flags       (merlin-lookup 'flags merlin-buffer-configuration))
-        (process-environment (cl-copy-list process-environment))
+        (process-environment
+         ;; for simplicity, we use a mere append here (leading to a
+         ;; duplicate binding), it does work because only the first
+         ;; occurrence is considered, one can check this by running
+         ;; (call-process "printenv" nil t)
+         (append (merlin-lookup 'env merlin-buffer-configuration)
+                 process-environment))
         (dot-merlin  (merlin-lookup 'dot-merlin merlin-buffer-configuration))
         ;; FIXME use logfile
         ;; (logfile     (or (merlin-lookup 'logfile merlin-buffer-configuration)
@@ -516,16 +526,6 @@ argument (lookup appropriate binary, setup logging, pass global settings)"
         (packages    (merlin--map-flatten (lambda (x) (cons "-I" x))
                                           merlin-buffer-packages-path))
         (filename    (buffer-file-name (buffer-base-buffer))))
-    ;; Update environment
-    (dolist (binding (merlin-lookup 'env merlin-buffer-configuration))
-      (let* ((equal-pos (string-match-p "=" binding))
-             (prefix (if equal-pos
-                       (substring binding 0 (1+ equal-pos))
-                       binding))
-             (is-prefix (lambda (x) (string-prefix-p prefix x))))
-        (setq process-environment (cl-delete-if is-prefix process-environment))
-        (when equal-pos
-          (setq process-environment (cons binding process-environment)))))
     ;; Compute verbosity
     (when (eq merlin-verbosity-context t)
       (setq merlin-verbosity-context (cons command args)))
@@ -565,16 +565,19 @@ argument (lookup appropriate binary, setup logging, pass global settings)"
 
 (defun merlin-call (command &rest args)
   "Execute a command and parse output: return an sexp on success or throw an error"
-  (let* ((binary (merlin-command))
-         (result (merlin--call-merlin command args)))
+  (let ((binary (merlin-command)) result)
     (condition-case err
-        (setq result (car (read-from-string result)))
-      (error
-        (merlin-client-logger binary command -1 "failure")
-        (error "merlin: error %s trying to parse answer: %s"
-               err result))
+        (progn
+          (setq result (merlin--call-merlin command args))
+          (condition-case err
+              (setq result (car (read-from-string result)))
+            (error
+             (merlin-client-logger binary command -1 "failure")
+             (error "merlin: error %s trying to parse answer: %s"
+                    err result))))
       (quit
-        (merlin-client-logger binary command -1 "interrupted")))
+       (merlin-client-logger binary command -1 "interrupted")
+       (signal (car err) (cdr err))))
     (let* ((notifications (cdr-safe (assoc 'notifications result)))
            (timing (cdr-safe (assoc 'timing result)))
            (class (cdr-safe (assoc 'class result)))
@@ -601,6 +604,35 @@ argument (lookup appropriate binary, setup logging, pass global settings)"
       (with-current-buffer buf
         (kill-local-variable 'merlin-buffer-configuration)
         (kill-local-variable 'merlin-erroneous-buffer)))))
+
+(defcustom merlin-stop-server-on-opam-switch t
+  "If t, stops the Merlin server before the opam switch changes.
+If the user changes the opam switch using `opam-switch-set-switch'
+or an `\"OPSW\"' menu from `opam-switch-mode', this option asks to
+stop the Merlin server process, so that the next Merlin command
+starts a new server, typically with a different Merlin version
+from a different opam switch.
+
+See https://github.com/ProofGeneral/opam-switch-mode
+
+Note: `opam-switch-mode' triggers automatic changes for `exec-path' and
+`process-environment', which are useful to find the `\"ocamlmerlin\"'
+binary (its filename can be overriden in `merlin-command') and the
+binary of Merlin's subprocesses, in the ambient opam switch."
+  :type 'boolean)
+
+(defun merlin--stop-server-on-opam-switch ()
+  "Stop the Merlin server before the opam switch changes.
+This function is for the `opam-switch-mode' hook
+`opam-switch-before-change-opam-switch-hook', which runs just
+before the user changes the opam switch through `opam-switch-mode'."
+  (when (and merlin-mode merlin-stop-server-on-opam-switch)
+    (condition-case _sig
+        (merlin-stop-server)
+      (t (message "Info: (merlin-stop-server) failed in the previous opam switch")))))
+
+(add-hook 'opam-switch-before-change-opam-switch-hook
+          #'merlin--stop-server-on-opam-switch t)
 
 ;;;;;;;;;;;;;;;;;;;;
 ;; FILE SWITCHING ;;
@@ -1455,19 +1487,30 @@ strictly within, or nil if there is no such element."
         (`(,result) (insert-choice result))
         (results (insert-choice (completing-read "Constructor: " results nil t)))))))
 
-(defun merlin--construct-point (point)
-  "Execute a construct at POINT."
+(defun merlin--construct-point (with-local-values point)
+  "Execute a construct at POINT.
+
+If WITH-LOCAL-VALUES is non-nil, pass \"-with-values local\" to
+include local values in the candidate list."
   (when-let ((result (merlin-call "construct"
-                                  "-position" (merlin-unmake-point point))))
+                                  "-position" (merlin-unmake-point point)
+                                  "-with-values"
+                                  (if (or with-local-values merlin-construct-with-local-values)
+                                      "local"
+                                    "null"))))
     (let* ((loc   (car result))
            (start (cdr (assoc 'start loc)))
            (stop  (cdr (assoc 'end loc))))
       (merlin--construct-complete start stop (cadr result)))))
 
-(defun merlin-construct ()
-  "Construct over the current hole."
-  (interactive)
-  (merlin--construct-point (point)))
+(defun merlin-construct (&optional arg)
+  "Construct over the current hole.
+
+With prefix ARG, include local values and not just constructors.
+This is like temporarily setting
+`merlin-construct-with-local-values' non-nil."
+  (interactive "P")
+  (merlin--construct-point arg (point)))
 
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
@@ -1890,7 +1933,8 @@ Empty string defaults to jumping to all these."
             (merlin-lookup 'do-not-cache-config merlin-buffer-configuration))
     (setq merlin-buffer-configuration (merlin--configuration)))
 
-  (let ((command (merlin-lookup 'command merlin-buffer-configuration)))
+  (let ((command (merlin-lookup 'command merlin-buffer-configuration))
+        bin-path)
     (unless command
       (setq
        command
@@ -1901,7 +1945,7 @@ Empty string defaults to jumping to all these."
          (with-temp-buffer
            (if (eq (call-process-shell-command
                     "opam var bin" nil (current-buffer) nil) 0)
-               (let ((bin-path
+               (let ((bin-dir
                       (replace-regexp-in-string "\n$" "" (buffer-string))))
                  ;; the opam bin dir needs to be on the path, so if merlin
                  ;; calls out to sub binaries (e.g. ocamlmerlin-reason), the
@@ -1910,21 +1954,17 @@ Empty string defaults to jumping to all these."
 
                  ;; this was originally done via `opam exec' but that does not
                  ;; work for opam 1, and added a performance hit
-                 (setq merlin-opam-bin-path (list (concat "PATH=" bin-path)))
-                 (concat bin-path "/ocamlmerlin"))
-
+                 (setq bin-path (list (concat "PATH=" bin-dir))))
              ;; best effort if opam is not available, lookup for the binary in
              ;; the existing env
-             (progn
-               (message "merlin-command: opam var failed (%S)"
-                        (buffer-string))
-               "ocamlmerlin"))))))
-
+             (message "merlin-command: opam var failed (%S)"
+                      (buffer-string)))
+           "ocamlmerlin"))))
       ;; cache command in merlin-buffer configuration to avoid having to shell
       ;; out to `opam` each time.
       (push (cons 'command command) merlin-buffer-configuration)
-      (when merlin-opam-bin-path
-        (push (cons 'env merlin-opam-bin-path) merlin-buffer-configuration)))
+      (when bin-path
+        (push (cons 'env bin-path) merlin-buffer-configuration)))
 
     command))