+++ src/ocaml/merlin_specific/browse_raw.ml
@@ -71,6 +71,7 @@ type node =
   | Class_declaration        of class_declaration
   | Class_description        of class_description
   | Class_type_declaration   of class_type_declaration
+  | Binding_op               of binding_op
 
   | Include_description      of include_description
   | Include_declaration      of include_declaration
@@ -111,6 +112,7 @@ let node_update_env env0 = function
   | Class_type_declaration  _ | Class_type_field        _
   | Include_description     _ | Include_declaration     _
   | Open_description        _ | Open_declaration        _
+  | Binding_op              _
     -> env0
 
 let node_real_loc loc0 = function
@@ -143,6 +145,7 @@ let node_real_loc loc0 = function
   | Include_declaration     {incl_loc = loc}
   | Open_description        {open_loc = loc}
   | Open_declaration        {open_loc = loc}
+  | Binding_op              {bop_op_name = {loc}}
     -> loc
   | Module_type_declaration_name {mtd_name = loc}
     -> loc.Location.loc
@@ -273,8 +276,8 @@ let of_constructor_arguments = function
   | Cstr_tuple cts -> list_fold of_core_type cts
   | Cstr_record lbls -> list_fold of_label_declaration lbls
 
-let of_bop { bop_op_path = _; bop_op_val = _; bop_exp; _ } =
-  of_expression bop_exp
+let of_bop ({ bop_exp; _ } as bop) =
+  app (Binding_op bop) ** of_expression bop_exp
 
 let of_record_field obj loc lbl =
   fun env (f : _ f0) acc ->
@@ -370,9 +373,17 @@ let of_expression_desc loc = function
   | Texp_unreachable | Texp_extension_constructor _ ->
     id_fold
   | Texp_letop { let_; ands; body; _ } ->
-    of_bop let_ **
-    list_fold of_bop ands **
-    of_case body
+    let rec flatten_patterns acc pat =
+      match pat.pat_desc with
+      | Tpat_tuple [ tuple; pat ] ->
+           flatten_patterns (pat :: acc) tuple
+      | _ -> List.rev (pat :: acc)
+    in
+    let bindops = let_ :: ands in
+    let patterns = flatten_patterns [] body.c_lhs in
+    let of_letop (pat, bindop) = of_bop bindop ** of_pattern pat in
+    list_fold of_letop (List.combine patterns bindops) **
+    of_expression body.c_rhs
   | Texp_open (od, e) ->
     app (Module_expr od.open_expr) ** of_expression e
 
@@ -673,6 +684,8 @@ let of_node = function
     of_module_expr i.incl_mod
   | Include_description i ->
     of_module_type i.incl_mod
+  | Binding_op { bop_exp=_ } ->
+    id_fold
 
 let fold_node f env node acc =
   of_node node env f acc
@@ -719,6 +732,7 @@ let string_of_node = function
   | Class_declaration       _ -> "class_declaration"
   | Class_description       _ -> "class_description"
   | Class_type_declaration  _ -> "class_type_declaration"
+  | Binding_op              _ -> "binding_op"
   | Method_call             _ -> "method_call"
   | Record_field            _ -> "record_field"
   | Module_binding_name     _ -> "module_binding_name"
@@ -919,6 +933,9 @@ let node_is_constructor = function
     Some {loc with Location.txt = `Description desc}
   | Pattern {pat_desc = Tpat_construct (loc, desc, _, _)} ->
     Some {loc with Location.txt = `Description desc}
+  | Extension_constructor ext_cons ->
+    Some { Location.loc = ext_cons.ext_loc;
+           txt = `Extension_constructor ext_cons}
   | _ -> None
 
 let node_of_binary_part env part =