+++ src/ocaml/merlin_specific/browse_raw.mli
@@ -84,6 +84,7 @@ type node =
   | Class_declaration        of class_declaration
   | Class_description        of class_description
   | Class_type_declaration   of class_type_declaration
+  | Binding_op               of binding_op
 
   | Include_description      of include_description
   | Include_declaration      of include_declaration
@@ -115,7 +116,9 @@ val node_paths_and_longident : node -> (Path.t Location.loc * Longident.t) list
 
 val node_is_constructor : node ->
   [ `Description of Types.constructor_description
-  | `Declaration of Typedtree.constructor_declaration ] Location.loc option
+  | `Declaration of Typedtree.constructor_declaration
+  | `Extension_constructor of Typedtree.extension_constructor ]
+    Location.loc option
 
 val node_of_binary_part : Env.t -> Cmt_format.binary_part -> node